+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### 16Apr2019
 - Added synapse location to `util\network\getLinkedNeurons.m` and edited the output parsing. If there's only a single output, all arguments will be returned as a table.
+- Updated functions using `util\network\getLinkedNeurons.m` to be compatible with new output
 
 ### 14Apr2019
 - Added function for returning renders in a figure (`util\renders\renderWhos.m`)